Added error checking for: Name North/south
blank lines after user input
clear console

frost bolt (will add AP)
mana shield (will give HP)
sout (check , and . )

add explanation of the skills for Warror Princess

remove useItem()

Console.clear()
X attack X
Console.pause()
X attack X

Game class -->Game controller
Game controller --> main
