+++ src/com/game/characters/Wizard.java
@@ -21,8 +21,8 @@ public class Wizard extends Heroes {
     public void useSpecialAbility() {
         while(true) {
             System.out.println("What Wizard skill you would like to use?");
-            System.out.println("1: Frost Bolt");
-            System.out.println("2. Mana Shield");
+            System.out.println("1: Frost Bolt: Will increase your attack damage");
+            System.out.println("2. Mana Shield: Will cast a shield that absorbs damage");
             System.out.println("Wizard mana = "+ getMp());
             Scanner scanner = new Scanner(System.in);
             String userInput = scanner.nextLine();
@@ -39,7 +39,7 @@ public class Wizard extends Heroes {
     }
 
     private void frostBolt() {
-        System.out.println("The Wizard casts the spell: frost bolt, the damage increases by 150%.");
+        System.out.println("The Wizard casts the spell: [Frost Bolt], the damage increases by 150%.");
         int frostBoltDamage = (int)(getAttackPower() *2.5);
         setAttackPower(frostBoltDamage);
         if (mp > 20) {
@@ -50,7 +50,7 @@ public class Wizard extends Heroes {
     }
 
     void manaShield() {
-        System.out.println("The Wizard casts the spell: manaShield. The wizard gained a shield");
+        System.out.println("The Wizard casts the spell: [Mana Shield], the wizard gained a shield");
         int shield = this.getMp();
         if (mp > 30) {
             hp = hp+shield;
